refactor(api): improve readability and structure in data and analysis services
